githubapi & newsapi moved to modules
