+++ src/js/modules/newsapi.js
